chore: simplify tests
